+++ Debug/Matrix.log
@@ -1,0 +1,22 @@
+﻿  Matrix.cpp
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(33,17): warning C4251: IndexError::ind_error: class "std::basic_string<char,std::char_traits<char>,std::allocator<char>>" должен иметь dll-интерфейс для использования клиентами class "IndexError"
+D:\VisualStudio\IDE\VC\Tools\MSVC\14.29.30133\include\xstring(4905): message : см. объявление "std::basic_string<char,std::char_traits<char>,std::allocator<char>>"
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(47,17): warning C4251: MatrixError::mtr_error: class "std::basic_string<char,std::char_traits<char>,std::allocator<char>>" должен иметь dll-интерфейс для использования клиентами class "MatrixError"
+D:\VisualStudio\IDE\VC\Tools\MSVC\14.29.30133\include\xstring(4905): message : см. объявление "std::basic_string<char,std::char_traits<char>,std::allocator<char>>"
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(64,17): warning C4251: FileError::fl_error: class "std::basic_string<char,std::char_traits<char>,std::allocator<char>>" должен иметь dll-интерфейс для использования клиентами class "FileError"
+D:\VisualStudio\IDE\VC\Tools\MSVC\14.29.30133\include\xstring(4905): message : см. объявление "std::basic_string<char,std::char_traits<char>,std::allocator<char>>"
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(88,38): warning C4251: Matrix::values: class "std::vector<std::vector<double,std::allocator<double>>,std::allocator<std::vector<double,std::allocator<double>>>>" должен иметь dll-интерфейс для использования клиентами class "Matrix"
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(88): message : см. объявление "std::vector<std::vector<double,std::allocator<double>>,std::allocator<std::vector<double,std::allocator<double>>>>"
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(118,27): warning C4018: <: несоответствие типов со знаком и без знака
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(131,24): warning C4244: =: преобразование "double" в "int", возможна потеря данных
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(132,24): warning C4244: =: преобразование "double" в "int", возможна потеря данных
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(814,27): warning C4018: <: несоответствие типов со знаком и без знака
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(1296,27): warning C4018: <: несоответствие типов со знаком и без знака
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(1310,27): warning C4018: <: несоответствие типов со знаком и без знака
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(1312,31): warning C4018: <: несоответствие типов со знаком и без знака
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(1328,27): warning C4018: <: несоответствие типов со знаком и без знака
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(1342,27): warning C4018: <: несоответствие типов со знаком и без знака
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(1347,27): warning C4018: <: несоответствие типов со знаком и без знака
+C:\Users\Кирилл\source\repos\Matrix\Matrix\Matrix.h(1348,31): warning C4018: <: несоответствие типов со знаком и без знака
+     Создается библиотека C:\Users\Кирилл\source\repos\Matrix\Debug\Matrix.lib и объект C:\Users\Кирилл\source\repos\Matrix\Debug\Matrix.exp
+  Matrix.vcxproj -> C:\Users\Кирилл\source\repos\Matrix\Debug\Matrix.dll
