+++ Debug/Matrix.tlog/Matrix.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.29.30133:VCServicingVersionATL=14.29.30136:VCServicingVersionCrtHeaders=14.29.30136:VCServicingVersionCompilers=14.29.30136:TargetPlatformVersion=10.0.18362.0:
+Debug|Win32|C:\Users\Кирилл\source\repos\Matrix\|
