+++ framework.h
@@ -1,0 +1,5 @@
+﻿#pragma once
+
+#define WIN32_LEAN_AND_MEAN             // Исключите редко используемые компоненты из заголовков Windows
+// Файлы заголовков Windows
+#include <windows.h>
